moved variables to common value and adressed other comments
