+++ src/main/java/org/opensearch/flowframework/model/State.java
@@ -15,5 +15,5 @@ public enum State {
     NOT_STARTED,
     PROVISIONING,
     FAILED,
-    READY
+    COMPLETED
 }
